+++ CHANGELOG.md
@@ -2,6 +2,21 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="3.12.0"></a>
+# [3.12.0](https://github.com/NetanelBasal/spectator/compare/v3.11.1...v3.12.0) (2019-08-15)
+
+
+### Bug Fixes
+
+* **mock:** fix regression caused by mocked properties ([#154](https://github.com/NetanelBasal/spectator/issues/154)) ([0bfc287](https://github.com/NetanelBasal/spectator/commit/0bfc287))
+
+
+### Features
+
+* **spectator:** use spectator element as default selector ([#155](https://github.com/NetanelBasal/spectator/issues/155)) ([6b05f76](https://github.com/NetanelBasal/spectator/commit/6b05f76))
+
+
+
 <a name="3.11.1"></a>
 ## [3.11.1](https://github.com/NetanelBasal/spectator/compare/v3.11.0...v3.11.1) (2019-08-08)